+++ src/main/java/fuzs/goldenagecombat/handler/ClassicCombatHandler.java
@@ -8,6 +8,7 @@ import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.attributes.Attribute;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.player.Player;
@@ -19,6 +20,9 @@ import net.minecraftforge.event.entity.ProjectileImpactEvent;
 import net.minecraftforge.event.entity.living.LivingEntityUseItemEvent;
 import net.minecraftforge.event.entity.player.AttackEntityEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+
+import java.util.function.BiConsumer;
+import java.util.function.Consumer;
 
 public class ClassicCombatHandler {
     @SubscribeEvent
@@ -32,31 +36,36 @@ public class ClassicCombatHandler {
     @SubscribeEvent
     public void onItemAttributeModifier(final ItemAttributeModifierEvent evt) {
         if (!GoldenAgeCombat.CONFIG.server().classic.oldAttackDamage) return;
+        if (evt.getSlotType() != EquipmentSlot.MAINHAND) return;
         ItemStack stack = evt.getItemStack();
-        if (!GoldenAgeCombat.CONFIG.server().classic.attackDamageBlacklist.contains(stack.getItem())) {
-            if (stack.getItem() instanceof TieredItem && evt.getSlotType() == EquipmentSlot.MAINHAND) {
-                // don't change items whose attributes have already been changed via the nbt tag
-                if (!stack.hasTag() || !stack.getTag().contains("AttributeModifiers", 9)) {
-                    // always one less to account for base value of 1.0
-                    if (stack.getItem() instanceof SwordItem) {
-                        this.replaceDamageAttribute(evt, (TieredItem) stack.getItem(), 4.0F);
-                    } else if (stack.getItem() instanceof AxeItem) {
-                        this.replaceDamageAttribute(evt, (TieredItem) stack.getItem(), 3.0F);
-                    } else if (stack.getItem() instanceof PickaxeItem) {
-                        this.replaceDamageAttribute(evt, (TieredItem) stack.getItem(), 2.0F);
-                    } else if (stack.getItem() instanceof ShovelItem) {
-                        this.replaceDamageAttribute(evt, (TieredItem) stack.getItem(), 1.0F);
-                    } else if (stack.getItem() instanceof HoeItem) {
-                        this.replaceDamageAttribute(evt, (TieredItem) stack.getItem(), 0.0F);
-                    }
+        // don't change items whose attributes have already been changed via the nbt tag
+        if (!stack.hasTag() || !stack.getTag().contains("AttributeModifiers", 9)) {
+            if (GoldenAgeCombat.CONFIG.server().classic.attackDamageOverrides.containsKey(stack.getItem())) {
+                this.replaceDamageAttribute(evt::removeAttribute, evt::addModifier, GoldenAgeCombat.CONFIG.server().classic.attackDamageOverrides.get(stack.getItem()));
+            } else if (stack.getItem() instanceof TieredItem) {
+                // always one less to account for base value of 1.0
+                if (stack.getItem() instanceof SwordItem) {
+                    this.replaceDamageAttribute(evt::removeAttribute, evt::addModifier, (TieredItem) stack.getItem(), 4.0);
+                } else if (stack.getItem() instanceof AxeItem) {
+                    this.replaceDamageAttribute(evt::removeAttribute, evt::addModifier, (TieredItem) stack.getItem(), 3.0);
+                } else if (stack.getItem() instanceof PickaxeItem) {
+                    this.replaceDamageAttribute(evt::removeAttribute, evt::addModifier, (TieredItem) stack.getItem(), 2.0);
+                } else if (stack.getItem() instanceof ShovelItem) {
+                    this.replaceDamageAttribute(evt::removeAttribute, evt::addModifier, (TieredItem) stack.getItem(), 1.0);
+                } else if (stack.getItem() instanceof HoeItem) {
+                    this.replaceDamageAttribute(evt::removeAttribute, evt::addModifier, (TieredItem) stack.getItem(), 0.0);
                 }
             }
         }
     }
 
-    private void replaceDamageAttribute(ItemAttributeModifierEvent evt, TieredItem item, float damageBonus) {
-        evt.removeAttribute(Attributes.ATTACK_DAMAGE);
-        evt.addModifier(Attributes.ATTACK_DAMAGE, new AttributeModifier(ItemAccessor.getBaseAttackDamageUUID(), new ResourceLocation(GoldenAgeCombat.MOD_ID, "attack_damage_modifier").toString(), item.getTier().getAttackDamageBonus() + damageBonus, AttributeModifier.Operation.ADDITION));
+    private void replaceDamageAttribute(Consumer<Attribute> removeAttribute, BiConsumer<Attribute, AttributeModifier> addModifier, TieredItem item, double damageBonus) {
+        this.replaceDamageAttribute(removeAttribute, addModifier, item.getTier().getAttackDamageBonus() + damageBonus);
+    }
+
+    private void replaceDamageAttribute(Consumer<Attribute> removeAttribute, BiConsumer<Attribute, AttributeModifier> addModifier, double newValue) {
+        removeAttribute.accept(Attributes.ATTACK_DAMAGE);
+        addModifier.accept(Attributes.ATTACK_DAMAGE, new AttributeModifier(ItemAccessor.getBaseAttackDamageUUID(), new ResourceLocation(GoldenAgeCombat.MOD_ID, "attack_damage_modifier").toString(), newValue, AttributeModifier.Operation.ADDITION));
     }
 
     @SubscribeEvent