+++ src/main/java/fuzs/goldenagecombat/handler/CombatAdjustmentsHandler.java
@@ -10,7 +10,7 @@ public class CombatAdjustmentsHandler {
     @SubscribeEvent
     public void onCriticalHit(final CriticalHitEvent evt) {
         // prevent sweeping from taking effect unless the enchantment is in place, onGround flag is reset next tick anyways
-        if (GoldenAgeCombat.CONFIG.server().adjustments.sweepingRequired && EnchantmentHelper.getSweepingDamageRatio(evt.getPlayer()) == 0.0F) {
+        if (GoldenAgeCombat.CONFIG.server().adjustments.sweepingRequired && EnchantmentHelper.getSweepingDamageRatio(evt.getPlayer()) == 0.0F || GoldenAgeCombat.CONFIG.server().adjustments.noSneakSweeping && evt.getPlayer().isShiftKeyDown()) {
             evt.getPlayer().setOnGround(false);
         }
     }