+++ src/main/java/fuzs/goldenagecombat/handler/SwordBlockingHandler.java
@@ -1,6 +1,7 @@
 package fuzs.goldenagecombat.handler;
 
 import fuzs.goldenagecombat.GoldenAgeCombat;
+import fuzs.goldenagecombat.registry.ModRegistry;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.damagesource.DamageSource;
@@ -65,12 +66,12 @@ public class SwordBlockingHandler {
     public static boolean canItemStackBlock(ItemStack stack) {
         if (!GoldenAgeCombat.CONFIG.server().blocking.allowBlocking) return false;
         Item item = stack.getItem();
-        if (GoldenAgeCombat.CONFIG.server().blocking.swordBlockingExclusions.contains(item)) {
+        if (GoldenAgeCombat.CONFIG.server().blocking.swordBlockingExclusions.contains(item) || stack.is(ModRegistry.SWORD_BLOCKING_EXCLUSIONS_TAG)) {
             return false;
         } else if (item instanceof SwordItem) {
             return true;
         } else {
-            return GoldenAgeCombat.CONFIG.server().blocking.swordBlockingInclusions.contains(item);
+            return GoldenAgeCombat.CONFIG.server().blocking.swordBlockingInclusions.contains(item) || stack.is(ModRegistry.SWORD_BLOCKING_INCLUSIONS_TAG);
         }
     }
 }